requirements.txt
aiopulsar-py
fastapi
uvicorn
pulsar-client
pulsar-client[avro]